backend/requirements.txt
fastapi[all]
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]
